back_end/main-master/requirements.txt
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
pytz==2020.5
sqlparse==0.4.1
psycopg2-binary==2.8.6
django-cors-headers==3.7.0
requests==2.25.1
locust==1.4.4

